FIxed accuracy

+++ FileParser.py
@@ -75,12 +75,12 @@ class FileParser:
         for i, file in enumerate(self.files):
 
             # Load images from .npy file
-            images = np.load(f"data/{file}").astype('float64') / 255.
+            images = np.load(f"data/{file}").astype('float32') / 255.
             # Reshape that each image will be in a 1X28X28 format
             images = images.reshape(-1, 1, 28, 28)
 
             # binary array that represents the location of the file in self.files
-            y = np.zeros(amount_of_files).astype('float64')
+            y = np.zeros(amount_of_files).astype('float32')
             y[i] = 1
 
             # Indices for storing data
